+++ src/components/ui/Buttons/GenderSelectButton.tsx
@@ -1,30 +1,32 @@
 interface GenderSelectButtonProps {
-  gender: string
+  genderId: number
+  genderValue: string
   selectedGender: string
   onChange: (gender: string) => void
 }
 
-export const GenderSelectButton: React.FC<GenderSelectButtonProps> = ({
-  gender,
+export function GenderSelectButton({
+  genderId,
+  genderValue,
   selectedGender,
   onChange,
-}) => {
-  const isSelected = selectedGender === gender
+}: GenderSelectButtonProps) {
+  const isSelected = selectedGender === genderValue
   return (
     <label
-      className={`flex justify-center items-center px-3 py-1.5 border border-solid rounded-[100px] ${
+      className={`flex justify-center items-center px-3 py-1.5 border border-solid rounded-full ${
         isSelected ? "bg-black text-white" : "border-zinc-500 text-black"
       }`}
     >
       <input
         type="radio"
         name="gender"
-        value={gender}
-        onChange={() => onChange(gender)}
+        value={genderId}
+        onChange={() => onChange(genderValue)}
         checked={isSelected}
         className="sr-only"
       />
-      <span>{gender}</span>
+      <span>{genderValue}</span>
     </label>
   )
 }
